+++ mycodo/mycodo_flask/forms/forms_settings.py
@@ -431,6 +431,8 @@ class SettingsDiagnostic(FlaskForm):
     delete_settings_database = SubmitField(lazy_gettext('Delete Settings Database'))
     delete_file_dependency = SubmitField(lazy_gettext('Delete File') + ': .dependency')
     delete_file_upgrade = SubmitField(lazy_gettext('Delete File') + ': .upgrade')
+    recreate_influxdb_db_1 = SubmitField('Recreate InfluxDB 1.x Database')
+    recreate_influxdb_db_2 = SubmitField('Recreate InfluxDB 2.x Database')
     reset_email_counter = SubmitField(lazy_gettext('Reset Email Counter'))
     install_dependencies = SubmitField(lazy_gettext('Install Dependencies'))
     regenerate_widget_html = SubmitField(lazy_gettext('Regenerate Widget HTML'))